Refactor of Movement Controllers
Refactoring the movement controllers to get the duplicate, networking
aware stuff out of the movement controllers.

Instead, we will have another class that will act as a mediator for
commands and will be network aware. See MovementHandler.cs for more
details.

NOTE: This commit breaks networking. Not sure if I want to stick with
Unity networking crap or switch all networking to SmartFox.